fix: update data types in session reports and AI services
- Changed the data type of `quotes_removed` in the session report model from Numeric to Integer for consistency.
- Updated the `build_translator` and `build_transcriber` functions to return specific types (Translator and Transcriber) instead of generic objects.
- Renamed the `translate` task to `translation` in various locations to avoid naming conflicts with the string method.
- Adjusted the metering and routing logic to reflect the new task name and updated call limits accordingly.
- Ensured that the handling of `quotes_removed` in the IeltsService correctly reflects the updated data type.

+++ backend/alembic/versions/0017_session_reports.py
@@ -51,7 +51,7 @@ def upgrade() -> None:
         sa.Column("drills", sa.JSON(), nullable=True),
         sa.Column("metrics", sa.JSON(), nullable=True),
         sa.Column("model", sa.String(length=48), nullable=False, server_default=""),
-        sa.Column("quotes_removed", sa.Numeric(4, 0), server_default="0", nullable=False),
+        sa.Column("quotes_removed", sa.Integer(), server_default="0", nullable=False),
         sa.Column(
             "created_at", sa.DateTime(timezone=True), server_default=sa.func.now(), nullable=False
         ),
